+++ RELEASES.md
@@ -82,6 +82,7 @@ Upcoming
     - Bugfix pipe description being showed from other configuration when both configurations contain the same adapter and pipe name
     - Show resources from schema, wsdl, fileName and schemaLocation attributes too in pipe description
 - Add LarvaPipe
+- Fix sub-resources eTag handling for ApiListener